lectures becomes positives

+++ billing/models.py
@@ -4,7 +4,7 @@ from django.urls import reverse
 
 class Lecture(models.Model):
     day = models.DateField()
-    lecture = models.FloatField()
+    lecture = models.PositiveIntegerField()
 
     def get_absolute_url(self):
         return reverse('lecture_detail', kwargs={'lecture_id': self.pk})
